refactor(build): hoist ForgetPolicy to me-types (L0)
ForgetPolicy (config in) and PruneStats (data out) are the input/output
pair of the same operation but sat two layers apart: ForgetPolicy lived
in me-forget (L3) while PruneStats already lived in me-types (L0). The
split was never justified by an actual dependency -- ForgetPolicy is
#[derive(Debug, Clone)] over scalar/collection fields whose only
non-local references (Result, FactType) were already L0.

#981 needs ForgetPolicy at L0: it promotes the dream-cycle capability
bag into an me-traits (L0.5) trait whose signature includes
`async fn forget(&self, policy: &ForgetPolicy) -> Result<PruneStats>`.
An L0.5 trait cannot name an L3 type without creating the back-edge
ADR-0018's break #1 exists to prevent.

Moves ForgetPolicy (struct, Default impl, inherent methods, and the 18
unit tests) wholesale into me-types/src/types/forgetting.rs, deletes
the now-empty me-forget/src/types.rs, and collapses me-forget's
re-export to a single `pub use me_types::types::forgetting::{ForgetPolicy,
PruneStats};`. The facade's re-export chain is untouched -- it already
resolves through me-forget. Replaces two stale peer-relative doc
rationales ("lives here because it isn't a trait" / "L0 cannot link up
to L3") with a property-based one: no L3 dependency -> L0, colocated
with the PruneStats it is paired with.

Closes #985.
